requirements.txt
flask
mysql-connector-python
pandas
scikit-learn
joblib
openpyxl
werkzeug
numpy
